Service Layer Completed

+++ model/card.go
@@ -4,6 +4,6 @@ type Card struct {
 	CardNumber string `json:"cardNumber"`
 	CardHolder string `json:"cardHolder"`
 	PinHash    string `json:"pinHash"`
-	Balance    int    `json:"balance"`
+	Balance    int64  `json:"balance"`
 	Status     string `json:"status"`
 }
